feat(skip): add higher-order lettable version of skip
