+++ src/operator/skip.ts
@@ -1,7 +1,5 @@
-import { Operator } from '../Operator';
-import { Subscriber } from '../Subscriber';
 import { Observable } from '../Observable';
-import { TeardownLogic } from '../Subscription';
+import { skip as higherOrder } from '../operators/skip';
 
 /**
  * Returns an Observable that skips the first `count` items emitted by the source Observable.
@@ -15,33 +13,5 @@ import { TeardownLogic } from '../Subscription';
  * @owner Observable
  */
 export function skip<T>(this: Observable<T>, count: number): Observable<T> {
-  return this.lift(new SkipOperator(count));
-}
-
-class SkipOperator<T> implements Operator<T, T> {
-  constructor(private total: number) {
-  }
-
-  call(subscriber: Subscriber<T>, source: any): TeardownLogic {
-    return source.subscribe(new SkipSubscriber(subscriber, this.total));
-  }
-}
-
-/**
- * We need this JSDoc comment for affecting ESDoc.
- * @ignore
- * @extends {Ignored}
- */
-class SkipSubscriber<T> extends Subscriber<T> {
-  count: number = 0;
-
-  constructor(destination: Subscriber<T>, private total: number) {
-    super(destination);
-  }
-
-  protected _next(x: T) {
-    if (++this.count > this.total) {
-      this.destination.next(x);
-    }
-  }
+  return higherOrder(count)(this);
 }
